Remove security checks on name_get
name_get is more complex to securize than other api methods.
The reason is that it is called when initializing a form view with default values.

For example, if a user clicks on 'Create' in an inventory adjustment form.
Suppose the default location is not authorized for this user.
The client calls name_get to get the name of the location.
An error is raised without letting the user change the location.
This situation is blocking.

Also, name_get is not much relevant for security.
Seeing the name of a record rarely is a security concern, because in most cases,
the name does not contain sensible information.

One exception is the name_get of employee addresses. However, for these records,
a rule is already defined at the orm level. Therefore, it does not need to be blocked at the rpc level.