Improve mock_odoo_request
Use a MockOdooHttpRequest class.

This class supports 2 properties redirect and website that are
added programatically by some odoo modules.

Fix a bug with the session store. Previous code used the session store
directly instead of creating a new session.

Properly propagate the cursor, uid and context from the env
to the request.

yield the request object so that it can be used inside the test.
This is required in some advanced test cases with website_sale related
features.